+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-## 23.8.5
-* Fixed a possible bug when 'enableRemoteConfigAutomaticTriggers' is enabled, getting remote config values just after the initialization raises error.
+## xx.x.x
+* Mitigated a rare issue related to handling request headers
   
 ## 23.8.4
 * Mitigated an issue where the device ID was changed after exiting an AB experiment